refactor: move HsConfig to a standalone file
